Ajout class MD25.
Replacement du DEBUG by DEBUG_UDP, afin de rajouter plus tard un autre
debug via la carte d'extension.

Replacement des int64_t by int32_t pour les codeurs, pid, config. On
laisse les calculs en UO avec des int64_t.

Ajout du cas #ifdef "CODEUR_AVEC_RESET" dans codeursDirects >
getCounts(). 

Main > Ajout de la ecouteSeriePC() pour avoir un debug directement sur
le port USB(serie).
Main > Ajout de la fonction ecouteI2C() pour attendre une instruction
sur l'autre port i2c. avec #ifdef COM_I2C_ACTIVATE.
 
Ajout du parametre perform_on sur le commandManager et sur le
ConsignController.

+++ asserv/filtres/Pid/Pid.h
@@ -22,13 +22,13 @@ private:
     int64_t integrale;
 
     //Coeffs d'asservissement
-    int64_t kp, ki, kd; // Coeff venant de config.h
+    int32_t kp, ki, kd; // Coeff venant de config.h
 
     //Ratios & Saturation
     /* Coeffs venant de config.h */
     double outRatio; // Ratio de sortie
-    int64_t maxIntegral; // Valeur maximale de l'intégrale
-    int64_t maxOutput; // Valeur maximale de la sortie
+    int32_t maxIntegral; // Valeur maximale de l'intégrale
+    int32_t maxOutput; // Valeur maximale de la sortie
 
 };
